+++ AdvProgProto1/Square.h
@@ -5,7 +5,7 @@
 
 class Square : public Shape, public Movable {
 private:
-	double edge;
+	float edge;
 public:
 
 	void calculateArea();
